Move bipartite adjacency map algorithms to correct file
