+++ test/Algebra/Graph/Test/Arbitrary.hs
@@ -30,7 +30,7 @@ import qualified Algebra.Graph.Acyclic.AdjacencyMap                        as AA
 import qualified Algebra.Graph.AdjacencyIntMap                             as AIM
 import qualified Algebra.Graph.AdjacencyMap                                as AM
 import qualified Algebra.Graph.Bipartite.Undirected.AdjacencyMap           as BAM
-import qualified Algebra.Graph.Bipartite.AdjacencyMap.Algorithm            as BAMA
+import qualified Algebra.Graph.Bipartite.Undirected.AdjacencyMap.Algorithm as BAMA
 import qualified Algebra.Graph.NonEmpty.AdjacencyMap                       as NAM
 import qualified Algebra.Graph.Class                                       as C
 import qualified Algebra.Graph.Labelled                                    as LG
